requirements.txt
# CORE DEPENDENCIES ONLY - For fast Railway deployment

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Payment Processing - ESSENTIAL
razorpay==1.4.2

# Authentication & Security
passlib[bcrypt]==1.7.4

# HTTP Client - For IP geolocation and social media APIs
httpx==0.25.2
requests==2.31.0

# Email - For confirmations
sendgrid==6.10.0
jinja2==3.1.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2